- добавил ендпоинт для notification service - отрефакторил контроллер - добавил метод для сервиса для получения юзер по айди
