+++ src/main/java/school/faang/user_service/dto/user/UserForNotificationDto.java
@@ -1,0 +1,14 @@
+package school.faang.user_service.dto.user;
+
+import school.faang.user_service.entity.contact.PreferredContact;
+import school.faang.user_service.entity.recommendation.Language;
+
+public record UserForNotificationDto(
+        long id,
+        String username,
+        String email,
+        String phone,
+        Language locale,
+        PreferredContact preference
+) {
+}
