+++ src/main/java/school/faang/user_service/entity/User.java
@@ -15,6 +15,7 @@ import school.faang.user_service.entity.goal.Goal;
 import school.faang.user_service.entity.goal.GoalInvitation;
 import school.faang.user_service.entity.event.Rating;
 import school.faang.user_service.entity.premium.Premium;
+import school.faang.user_service.entity.recommendation.Language;
 import school.faang.user_service.entity.recommendation.Recommendation;
 
 import java.time.LocalDateTime;
@@ -147,6 +148,10 @@ public class User {
 
     @Column(name = "banned")
     private boolean banned;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "locale", nullable = false)
+    private Language locale;
 
     public void removeOwnedEvent(Event event){
         ownedEvents.remove(event);
