+++ src/main/java/school/faang/user_service/entity/recommendation/Language.java
@@ -1,0 +1,15 @@
+package school.faang.user_service.entity.recommendation;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public enum Language {
+    EN("en"),
+    RU("ru"),
+    FR("fr"),
+    DE("de");
+
+    private final String tag;
+}
