+++ src/main/resources/db/changelog/changeset/user_V009_user_add_column_locale.sql
@@ -1,0 +1,2 @@
+ALTER TABLE users
+ADD COLUMN locale VARCHAR(10) DEFAULT 'en';
