+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -15,3 +15,5 @@ databaseChangeLog:
       file: db/changelog/changeset/user_V007__user_premium.sql
   - include:
       file: db/changelog/changeset/user_V008_user_add_column_banned.sql
+  - include:
+      file: db/changelog/changeset/user_V009_user_add_column_locale.sql
